Split up diffs and patches
diffs can be applied as-is.

patches must be massaged by the randomizer code before being applied.

i suppose in general another option would be to leave them combined
and have all of the functions in the patch namespace operate on the
entire list of patches; in that case, each function would be
responsible for either updating the patch or deleting it from the
list. this solution would also push logic out of the main patch
function, which is a plus.